refactor: add session manager interface class
This way we can split up the implementation in the future for different
windowing systems.
